// src/app/api-client/model/imageGenerationModelRecordInput.ts
/**
 * FastAPI
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ImageGenerationModelRecordInputRequirementsValue } from './imageGenerationModelRecordInputRequirementsValue';
import { ModelClassification } from './modelClassification';
import { GenericModelRecordConfig } from './genericModelRecordConfig';
import { GenericModelRecordMetadata } from './genericModelRecordMetadata';
import { FineTuneSeriesInfo } from './fineTuneSeriesInfo';
import { MODEL_REFERENCE_CATEGORY } from './mODELREFERENCECATEGORY';

/**
 * A model entry in the model reference.
 */
export interface ImageGenerationModelRecordInput {
  record_type?: MODEL_REFERENCE_CATEGORY;
  name: string;
  description?: string | null;
  version?: string | null;
  finetune_series?: FineTuneSeriesInfo | null;
  metadata?: GenericModelRecordMetadata;
  config?: GenericModelRecordConfig;
  model_classification?: ModelClassification;
  inpainting?: boolean | null;
  baseline: string;
  optimization?: string | null;
  tags?: Array<string> | null;
  showcases?: Array<string> | null;
  min_bridge_version?: number | null;
  trigger?: Array<string> | null;
  homepage?: string | null;
  nsfw: boolean;
  style?: string | null;
  requirements?: { [key: string]: ImageGenerationModelRecordInputRequirementsValue } | null;
  size_on_disk_bytes?: number | null;
}
export namespace ImageGenerationModelRecordInput {}